cadastre_middleware/requirements.txt
black==23.10.0
Flask==3.0.0
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0